Move dashboard-only components into dashboard
