+++ frontend/src/metabase/dashboard/components/AIDashboardAnalysisSidebar/AIDashboardAnalysisSidebar.unit.spec.tsx
