+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/get-supported-cards-for-subscriptions.tsx
